MyClub/src/main/java/com/clemson/model/TestSchool.java
package com.clemson.model;

public class TestSchool {
	private int testId;
	private int schoolId;

	// 外键关联项
	private String schoolName;

	public int getTestId() {
		return testId;
	}

	public void setTestId(int testId) {
		this.testId = testId;
	}

	public int getSchoolId() {
		return schoolId;
	}

	public void setSchoolId(int schoolId) {
		this.schoolId = schoolId;
	}
	
	//
	public String getSchoolName() {
		return schoolName;
	}
	public void setSchoolName(String schoolName) {
		this.schoolName = schoolName;
	}
}
